tools: make build of FLASK utils/policy explicit
This patch makes the generation of the FLASK utilities and the policy
explicit. The result is two true user facing options for XSM/FLASK. The
first being in the hypervisor that is available through Kconfig and the
second being in the tools stack with --{enable,disable}-flask. Most of
the distros build the hypervisor separate from the tools stack currently
so having two switches should not affect their workflow. It does drop
the two knobs that could affect the tools stack by removing FLASK_ENABLE
that was set at the top level and changes --{enable,disable}-xsmpolicy
to --{enable,disable}-flask to unify them.

This drops the two control knobs in the tools that affect how the FLASK
policy and tools are generated and unifies it under one. After this
change it should

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>