x86/mtrr: prefix fns with mtrr and drop static
For the functions that make up the interface to the MTRR code, drop the
static keyword and prefix them all with mtrr for improved clarity when
they're called outside this file. This also required adjusting or
providing function prototypes to make them callable. Future changes will
directly call these functions instead of using the indirect call through
the ops structure.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>