x86/mtrr: drop mtrr_if indirection
There can only ever be one mtrr_if now and that is the generic
implementation so instead of going through an indirect call change
everything to call the generic implementation directly. The is_cpu()
macro would result in the left side always being equal to
X86_VENDOR_INTEL for the generic implementation due to Intel having a
value of 0. The use_intel() macro was always true in the generic
implementation case as well. These macros are addressed in later
changes.  Removed some extraneous whitespace at the same time.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>