x86/EFI: fix build when using GNU Make 4.1
Since c/s eee5909e9d (x86/EFI: use less crude a way of generating the
build ID) builds have been broken when using GNU Make 4.1. This is
because there are no dependencies on buildid.o and as such GNU Make does
not build it. This adds a dependency so that it is built.

Note: This patch is different than I was going to originally submit, I
had used runtime.o but while testing this I saw Daniel Kiper's v14 of
his multiboot2 series come along and I modified my patch so that he
would not have to rebase.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>